pyproject.toml
[project]
name = "segmentacao-imagens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch==2.6.0",
    "torchvision==0.21.0",
    "torchaudio==2.6.0",
    "pillow",
    "numpy>=2.1.0",
    "typing",
    "scikit-image",
    "huggingface_hub",
    "transformers>=4.39.1",
    "gradio",
    "pysentimiento",
    "liqfit",
    "sentencepiece",
]
